Additional refactoring
